@@ -1,15 +1,12 @@
 module github.com/cerbos/protoc-gen-go-hashpb
 
-go 1.21
+go 1.23
 
-toolchain go1.23.3
-
-require (
-	github.com/cespare/xxhash/v2 v2.1.2
-	google.golang.org/protobuf v1.36.0
-)
+toolchain go1.24.1
 
 require (
-	github.com/google/go-cmp v0.5.6 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	github.com/cespare/xxhash/v2 v2.3.0
+	google.golang.org/protobuf v1.36.5
 )
+
+require github.com/google/go-cmp v0.7.0 // indirect
