@@ -1,7 +1,8 @@
 module github.com/cerbos/hashpb/tools
 
-go 1.22.0
-toolchain go1.23.4
+go 1.23.0
+
+toolchain go1.24.1
 
 require (
 	github.com/bufbuild/buf v1.30.0
